+++ src/relstorage/adapters/mysql/schema.py
@@ -28,19 +28,14 @@ class MySQLSchemaInstaller(AbstractSchemaInstaller):
 
     database_type = 'mysql'
 
-    def _to_native_str(self, value):
-        if not isinstance(value, str):
-            value = value.decode('ascii')
-        return value
-
     def get_database_name(self, cursor):
         cursor.execute("SELECT DATABASE()")
         for (name,) in cursor:
-            return self._to_native_str(name)
+            return self._metadata_to_native_str(name)
 
     def list_tables(self, cursor):
         cursor.execute("SHOW TABLES")
-        return [self._to_native_str(name)
+        return [self._metadata_to_native_str(name)
                 for (name,) in cursor.fetchall()]
 
     def list_sequences(self, cursor):
@@ -48,18 +43,15 @@ class MySQLSchemaInstaller(AbstractSchemaInstaller):
 
     def check_compatibility(self, cursor, tables):
         super(MySQLSchemaInstaller, self).check_compatibility(cursor, tables)
+        # TODO: Check more tables, like `transaction`
         stmt = "SHOW TABLE STATUS LIKE 'object_state'"
         cursor.execute(stmt)
-        for row in cursor:
-            for col_index, col in enumerate(cursor.description):
-                if col[0].lower() == 'engine':
-                    engine = row[col_index]
-                    if not isinstance(engine, str):
-                        engine = engine.decode('ascii')
-                    if engine.lower() != 'innodb':
-                        raise StorageError(
-                            "The object_state table must use the InnoDB "
-                            "engine, but it is using the %s engine." % engine)
+        for row in self._rows_as_dicts(cursor):
+            engine = self._metadata_to_native_str(row['engine'])
+            if engine.lower() != 'innodb':
+                raise StorageError(
+                    "The object_state table must use the InnoDB "
+                    "engine, but it is using the %s engine." % engine)
 
     def _create_commit_lock(self, cursor):
         return
@@ -67,20 +59,17 @@ class MySQLSchemaInstaller(AbstractSchemaInstaller):
     def _create_pack_lock(self, cursor):
         return
 
+    COLTYPE_BINARY_STRING = 'BLOB'
+    TRANSACTIONAL_TABLE_SUFFIX = 'ENGINE = InnoDB'
 
-    def _create_transaction(self, cursor):
-        if self.keep_history:
-            stmt = """
-            CREATE TABLE transaction (
-                tid         BIGINT NOT NULL PRIMARY KEY,
-                packed      BOOLEAN NOT NULL DEFAULT FALSE,
-                empty       BOOLEAN NOT NULL DEFAULT FALSE,
-                username    BLOB NOT NULL,
-                description BLOB NOT NULL,
-                extension   BLOB
-            ) ENGINE = InnoDB;
-            """
-            self.runner.run_script(cursor, stmt)
+    # As usual, MySQL has an annoying implementation of this and we
+    # have to re-specify *everything* about the column. MySQL 8 supports the
+    # simple 'RENAME ... TO ... syntax that everyone else does.
+    _rename_transaction_empty_stmt = (
+        "ALTER TABLE transaction CHANGE empty is_empty "
+        "BOOLEAN NOT NULL DEFAULT FALSE"
+    )
+
 
     def _create_new_oid(self, cursor):
         stmt = """