+++ CHANGES.rst
@@ -46,6 +46,22 @@
 - Make driver names in RelStorage configurations case-insensitive
   (e.g., 'MySQLdb' and 'mysqldb' are both valid). See :issue:`227`.
 
+- Rename the column ``transaction.empty`` to ``transaction.is_empty``
+  for compatibility with MySQL 8.0, where ``empty`` is now a reserved
+  word. The migration will happen automatically when a storage is
+  first opened, unless it is configured not to create the schema.
+
+  .. note:: This migration has not been tested for Oracle.
+
+  .. note:: You must run this migration *before* attempting to upgrade
+            a MySQL 5 database to MySQL 8. If you cannot run the
+            upgrade through opening the storage, the statement is
+            ``ALTER TABLE transaction CHANGE empty is_empty BOOLEAN
+            NOT NULL DEFAULT FALSE``.
+
+- Stop getting a warning about invalid optimizer syntax when packing a
+  MySQL database (especially with the PyMySQL driver). See
+  :issue:`163`.
 
 2.1.1 (2019-01-07)
 ==================